FileAPI/FileReader/test_securityerrors-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>FileReader SecurityError Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/FileAPI/#dfn-error-codes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<form name="upload">
    <input type="file" id="fileChooser"><br><input type="button" id="start" value="start">
</form>

<div>
    <p>Test steps:</p>
    <ol>
        <li>Download the <a href="support/file_test1.txt">file</a>.</li>
        <li>Select the file in the file inputbox.</li>
        <li>Change the file.</li>
        <li>Click the 'start' button.</li>
    </ol>
</div>

<div id="log"></div>

<script>

    var fileChooser = document.querySelector('#fileChooser');

    setup({explicit_done: true});
    setup({explicit_timeout: true});

    on_event(fileChooser, 'change', function() {

        async_test(function(t) {

            var reader = new FileReader();
            reader.readAsArrayBuffer(fileChooser.files[0]);

            reader.onloadend = t.step_func_done(function(event) {
                assert_equals(event.target.readyState, FileReader.DONE);
                assert_equals(reader.error, null);
            });

        }, 'FileReader.error should be null if there are no errors when reading');

    });

    on_event(document.querySelector('#start'), 'click', function() {

        async_test(function(t) {

            var reader = new FileReader();
            reader.readAsArrayBuffer(fileChooser.files[0]);

            reader.onloadend = t.step_func_done(function(event) {
                assert_equals(event.target.readyState, FileReader.DONE);
                assert_equals(reader.error.code, 18);
            });

        }, 'FileReader.error should be SECURITY_ERROR if the file has changed since the user selected it');

        done();

    });

</script>
